:hammer: View와 관련된 함수 분리
