+++ src/js/display-utils.js
@@ -1,0 +1,61 @@
+import { parseHTML, sections, state, resetGame } from "./index.js";
+
+const hideElement = (element) => {
+  return (element.style.display = "none");
+};
+
+export const showElement = (element) => {
+  return (element.style.display = "block");
+};
+
+export const resetView = (elementIdArray) => {
+  for (let elementId of elementIdArray) {
+    hideElement(sections[elementId]);
+  }
+};
+
+const showCarName = (carName) => {
+  return parseHTML(`<div class="car-player mr-2">${carName}</div>`);
+};
+
+const showTotalStep = () => {
+  return parseHTML(`<div class="forward-icon mt-2">⬇️️</div>`);
+};
+
+export const setResultView = () => {
+  sections[3].querySelector("div").innerHTML = "";
+
+  state.cars.forEach((car) => {
+    const resultDivString = `<div></div>`;
+    const resultDiv = parseHTML(resultDivString);
+
+    resultDiv.appendChild(showCarName(car.name));
+    for (let idx = 0; idx < car.totalStep; idx++) {
+      const step = showTotalStep();
+      resultDiv.appendChild(step);
+    }
+    sections[3].querySelector("div").append(resultDiv);
+  });
+};
+
+export const setWinnerView = (winners) => {
+  sections[4].innerHTML = "";
+  let winnerText = "";
+  if (winners.length === 1) {
+    winnerText = winners[0];
+  } else {
+    winnerText = winners.join(", ");
+  }
+
+  const winnerTemplateString = `<h2>🏆 최종 우승자: ${winnerText} 🏆</h2>`;
+  const winnerTemplate = parseHTML(winnerTemplateString);
+  const resetBtnString = `<div class="d-flex justify-center">
+                            <button type="button" class="btn btn-cyan">다시 시작하기</button>
+                          </div>`;
+  const resetBtn = parseHTML(resetBtnString);
+
+  sections[4].append(winnerTemplate);
+  sections[4].append(resetBtn);
+
+  resetGame();
+};
