+++ src/js/index.js
@@ -1,25 +1,18 @@
 import Car from "./Car.js";
-const parser = new DOMParser();
-const sections = document.getElementsByTagName("section");
-const state = {
+import {
+  showElement,
+  resetView,
+  setResultView,
+  setWinnerView,
+} from "./display-utils.js";
+
+export const sections = document.getElementsByTagName("section");
+export const state = {
   cars: [],
 };
 
-const hideElement = (element) => {
-  return (element.style.display = "none");
-};
-
-const showElement = (element) => {
-  return (element.style.display = "block");
-};
-
-const resetView = (elementIdArray) => {
-  for (let elementId of elementIdArray) {
-    hideElement(sections[elementId]);
-  }
-};
-
-const parseHTML = (html) => {
+export const parseHTML = (html) => {
+  const parser = new DOMParser();
   return parser.parseFromString(html, "text/html").body.firstElementChild;
 };
 
@@ -73,30 +66,6 @@ const playGame = () => {
   }
 };
 
-const showCarName = (carName) => {
-  return parseHTML(`<div class="car-player mr-2">${carName}</div>`);
-};
-
-const showTotalStep = () => {
-  return parseHTML(`<div class="forward-icon mt-2">⬇️️</div>`);
-};
-
-const setResultView = () => {
-  sections[3].querySelector("div").innerHTML = "";
-
-  state.cars.forEach((car) => {
-    const resultDivString = `<div></div>`;
-    const resultDiv = parseHTML(resultDivString);
-
-    resultDiv.appendChild(showCarName(car.name));
-    for (let idx = 0; idx < car.totalStep; idx++) {
-      const step = showTotalStep();
-      resultDiv.appendChild(step);
-    }
-    sections[3].querySelector("div").append(resultDiv);
-  });
-};
-
 const getWinner = () => {
   state.cars.sort((a, b) => {
     return b.totalStep - a.totalStep;
@@ -114,7 +83,7 @@ const getWinner = () => {
   });
 };
 
-const resetGame = () => {
+export const resetGame = () => {
   const carNamesBtn = document.getElementsByTagName("button")[0];
   const tryNumBtn = document.getElementsByTagName("button")[1];
   const resetBtn = document.getElementsByTagName("button")[2];
@@ -128,28 +97,6 @@ const resetGame = () => {
     tryNumBtn.disabled = false;
     carNamesBtn.disabled = false;
   });
-};
-
-const setWinnerView = (winners) => {
-  sections[4].innerHTML = "";
-  let winnerText = "";
-  if (winners.length === 1) {
-    winnerText = winners[0];
-  } else {
-    winnerText = winners.join(", ");
-  }
-
-  const winnerTemplateString = `<h2>🏆 최종 우승자: ${winnerText} 🏆</h2>`;
-  const winnerTemplate = parseHTML(winnerTemplateString);
-  const resetBtnString = `<div class="d-flex justify-center">
-                            <button type="button" class="btn btn-cyan">다시 시작하기</button>
-                          </div>`;
-  const resetBtn = parseHTML(resetBtnString);
-
-  sections[4].append(winnerTemplate);
-  sections[4].append(resetBtn);
-
-  resetGame();
 };
 
 const onClickedtryNumBtn = () => {
